LayoutTests/imported/w3c/web-platform-tests/html/dom/render-blocking/script-inserted-script-expected.txt

Some text

PASS Rendering is blocked before render-blocking resources are loaded
PASS Script-inserted render-blocking script is evaluated
